benerin tipe data di data class

+++ app/src/main/java/com/adit/bangkit/plagroid/ui/activities/AddProductActivity.kt
@@ -236,9 +236,9 @@ class AddProductActivity : BaseActivity(), View.OnClickListener {
             FirestoreClass().getCurrentUserID(),
             username,
             et_product_title.text.toString().trim { it <= ' ' },
-            et_product_price.text.toString().trim { it <= ' ' },
+            et_product_price.text.toString().trim { it <= ' ' }.toInt(),
             et_product_description.text.toString().trim { it <= ' ' },
-            et_product_quantity.text.toString().trim { it <= ' ' },
+            et_product_quantity.text.toString().trim { it <= ' ' }.toInt(),
             mProductImageURL
         )
